Migrate to asset pipeline Requires secuirty.cas.appServerName for travis
