+++ grails-app/assets/javascripts/sockets.js
@@ -1,0 +1,6 @@
+
+// Websockets asset, used for feedback during admin
+
+//= require sockjs.js
+//= require stomp.js
+
