--- grails-app/conf/ApplicationResources.groovy
@@ -1,7 +1,0 @@
-modules = {
-    sockets {
-        dependsOn 'bootstrap'
-        resource url: [dir: 'js', file: 'sockjs.js']
-        resource url: [dir: 'js', file: 'stomp.js']
-    }
-}
