+++ grails-app/views/import/collectory.gsp
@@ -4,7 +4,7 @@
   <title><g:message code="admin.import.collectory.label"/></title>
   <meta name="layout" content="${grailsApplication.config.skin.layout}"/>
     <meta name="breadcrumbParent" content="${createLink(controller:'admin', action:'index', absolute:true)},${message(code: 'breadcrumb.admin')}"/>
-    <r:require modules="sockets" />
+    <asset:javascript src="sockets"/>
 </head>
 <body>
 <div>
@@ -27,7 +27,7 @@
         </div>
     </div>
 
-    <r:script>
+    <asset:script type="text/javascript">
         function loadCollectoryInfo(){
             $.get("${createLink(controller:'import', action:'importCollectory')}", function( data ) {
               if(data.success){
@@ -39,9 +39,9 @@
               $('.import-info').removeClass('hide');
             });
         }
-    </r:script>
+    </asset:script>
 
-    <r:script>
+    <asset:script type="text/javascript">
         $(function() {
             var socket = new SockJS("${createLink(uri: '/stomp')}");
             var client = Stomp.over(socket);
@@ -51,7 +51,7 @@
                 });
             });
         });
-    </r:script>
+    </asset:script>
 </div>
 </body>
 </html>
