+++ grails-app/views/import/index.gsp
@@ -4,7 +4,7 @@
   <title><g:message code="admin.import.dwca.label"/></title>
   <meta name="layout" content="${grailsApplication.config.skin.layout}"/>
   <meta name="breadcrumbParent" content="${createLink(controller:'admin', action:'index', absolute:true)},${message(code: 'breadcrumb.admin')}"/>
-  <r:require modules="sockets" />
+    <asset:javascript src="sockets"/>
 </head>
 <body>
 <div>
@@ -53,7 +53,7 @@
         </div>
     </div>
 
-    <r:script>
+    <asset:script type="text/javascript">
 
         function loadDwCAFromDir(){
             loadDwCA($('#dwca_dir').val())
@@ -69,9 +69,9 @@
               $('.import-info').removeClass('hide');
             });
         }
-    </r:script>
+    </asset:script>
 
-    <r:script>
+    <asset:script type="text/javascript">
         $(function() {
             var socket = new SockJS("${createLink(uri: '/stomp')}");
             var client = Stomp.over(socket);
@@ -81,7 +81,7 @@
                 });
             });
         });
-    </r:script>
+    </asset:script>
 </div>
 </body>
 </html>
