+++ grails-app/views/import/specieslist.gsp
@@ -17,7 +17,7 @@
   <title><g:message code="admin.import.specieslist.label"/></title>
   <meta name="layout" content="${grailsApplication.config.skin.layout}"/>
     <meta name="breadcrumbParent" content="${createLink(controller:'admin', action:'index', absolute:true)},${message(code: 'breadcrumb.admin')}"/>
-    <r:require modules="sockets" />
+    <asset:javascript src="sockets"/>
     <style type="text/css">
         .progress {
             height: 10px !important;
@@ -61,7 +61,7 @@
         </div>
     </div>
 
-    <r:script>
+    <asset:script type="text/javascript">
         function loadConservationSpeciesLists(){
             $.get("${createLink(controller:'import', action:'importConservationSpeciesLists')}", function( data ) {
               if(data.success){
@@ -88,9 +88,9 @@
             });
         }
 
-    </r:script>
+    </asset:script>
 
-    <r:script>
+    <asset:script type="text/javascript">
         $(function() {
             var socket = new SockJS("${createLink(uri: '/stomp')}");
             var client = Stomp.over(socket);
@@ -116,7 +116,7 @@
                 });
             });
         });
-    </r:script>
+    </asset:script>
 </div>
 </body>
 </html>
