+++ grails-app/views/import/wordpress.gsp
@@ -17,7 +17,7 @@
   <title>WordPress Import</title>
   <meta name="layout" content="${grailsApplication.config.skin.layout}"/>
     <meta name="breadcrumbParent" content="${createLink(controller:'admin', action:'index', absolute:true)},${message(code: 'breadcrumb.admin')}"/>
-    <r:require modules="sockets" />
+    <asset:javascript src="sockets"/>
     <style type="text/css">
         .progress {
             height: 10px !important;
@@ -50,7 +50,7 @@
         </div>
     </div>
 
-    <r:script>
+    <asset:script type="text/javascript">
         function loadWordPressPages(){
             $.get("${createLink(controller:'import', action:'importWordPress')}", function( data ) {
               if(data.success){
@@ -63,9 +63,9 @@
               $('.progress').removeClass('hide');
             });
         }
-    </r:script>
+    </asset:script>
 
-    <r:script>
+    <asset:script type="text/javascript">
         $(function() {
             var socket = new SockJS("${createLink(uri: '/stomp')}");
             var client = Stomp.over(socket);
@@ -84,7 +84,7 @@
                 });
             });
         });
-    </r:script>
+    </asset:script>
 </div>
 </body>
 </html>
